+++ Changelog.md
@@ -3,6 +3,7 @@ AppleALC Changelog
 
 #### v1.0.1
 - boot-args presence is now optional
+- Added AD1988B codec support (by Rodion)
 
 #### v1.0.0
 - Initial release